+++ src/App.tsx
@@ -1,17 +1,17 @@
-import Navbar from "./components/layout//Navbar";
+import Navbar from "./components/Navbar";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 // import Hero from "./components/Hero";
-import About from "./components/pages/About";
-import Posts from "./components/pages/Posts";
+import About from "./components/layout/About";
+import Videos from "./components/layout/Videos";
 
 const App = () => {
   return (
     <>
-      <Navbar />
       <Router>
+        <Navbar />
         <Routes>
           <Route path="/" element={<About />} />
-          <Route path="/posts" element={<Posts />} />
+          <Route path="/videos" element={<Videos />} />
         </Routes>
       </Router>
     </>
