+++ src/components/layout/About.tsx
@@ -1,0 +1,13 @@
+// https://mantine.dev/core/timeline/
+import { Flex, Box, Image, VStack, Container } from "@chakra-ui/react";
+const About = () => {
+  return (
+    <Box pt={20}>
+      <Container align="center">
+        <Image src="/hi.png" boxSize="200px" />
+      </Container>
+    </Box>
+  );
+};
+
+export default About;
