+++ src/components/layout/Videos.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Videos = () => {
+  return <div>Videos</div>;
+};
+
+export default Videos;
