--- src/components/pages/About.tsx
@@ -1,6 +1,0 @@
-// https://mantine.dev/core/timeline/
-const About = () => {
-  return <div>About</div>;
-};
-
-export default About;
