--- src/components/pages/Posts.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Posts = () => {
-  return <div>Posts</div>;
-};
-
-export default Posts;
